v2.0.9 fixes and cli improvments for OBC
