+++ doc/backing-store-crd.md
@@ -5,7 +5,7 @@ BackingStore CRD represents a storage target to be used as underlying storage fo
 These storage targets are used to store deduped+compressed+encrypted chunks of data (encryption keys are stored separately).
 Backing-stores are referred to by name when defining [BucketClass](bucket-class-crd.md).
 
-Multiple types of backing-stores are currently supported: aws-s3, s3-compatible, google-cloud-storage, azure-blob.
+Multiple types of backing-stores are currently supported: aws-s3, s3-compatible, google-cloud-storage, azure-blob, pv-pool.
 Backing-store type pv-pool is not yet supported by the operator. Instead, the web management console must be used to administer pv-pool backing-stores.
 Adding support for a new type of backing-store is rather easy as it requires just GET/PUT key-value store, see [Backing-stores supported by NooBaa](https://github.com/noobaa/noobaa-core/tree/master/src/agent/block_store_services).
 
@@ -121,75 +121,29 @@ spec:
 ```
 
 #### PV-POOL type
-**Not yet implemented**
 
-Create NooBaa resources StatefulSet with PVC mounted in each pod. Each resource will connect to the NooBaa brain and provide the PV filesystem storage to be used for storing encrypted chunks of data.
-This action is supported from the NooBaa dashboard (Deploy Kubernetes Pool).
-It is possible to configure the number of pods to be used and their PV size.
-Here is an example of a StatefulSet with 3 pods and PV size of 30GB:
+Create NooBaa resources StatefulSet with PVC mounted in each pod. Each resource will connect to the NooBaa core and provide the PV filesystem storage to be used for storing encrypted chunks of data. It is possible to configure the number of pods to be used and their PV size.
+```shell
+noobaa -n noobaa backingstore create pv-pool bs --num-volumes 3 --pv-size-gb 32 --storage-class STORAGE-CLASS-NAME
+```
 ```yaml
----
-apiVersion: apps/v1
-kind: StatefulSet
+apiVersion: noobaa.io/v1alpha1
+kind: BackingStore
 metadata:
-  name: noobaa-agent
+  finalizers:
+  - noobaa.io/finalizer
   labels:
     app: noobaa
-    noobaa-module: noobaa-pool-impl
+  name: bs
+  namespace: noobaa
 spec:
-  selector:
-    matchLabels:
-      noobaa-module: noobaa-agent
-  serviceName: noobaa-agent
-  replicas: 3
-  updateStrategy:
-    type: RollingUpdate
-  template:
-    metadata:
-      labels:
-        app: noobaa
-        noobaa-module: noobaa-agent
-        noobaa-s3: "true"
-    spec:
-      containers:
-        - name: noobaa-agent
-          resources:
-            requests:
-              cpu: "100m"
-              memory: "500Mi"
-            limits:
-              cpu: "2"
-              memory: "2Gi"
-          env:
-            - name: CONTAINER_PLATFORM
-              value: KUBERNETES
-            - name: AGENT_CONFIG
-              value: "AGENT_CONFIG_VALUE"
-            - name: ENDPOINT_PORT
-              value: "6001"
-            - name: ENDPOINT_SSL_PORT
-              value: "6443"
-          command: ["/noobaa_init_files/noobaa_init.sh", "agent"]
-          ports:
-            - containerPort: 60101
-          volumeMounts:
-            - name: noobaastorage
-              mountPath: /noobaa_storage
-            - name: tmp-logs-vol
-              mountPath: /usr/local/noobaa/logs
-      volumes:
-        - name: tmp-logs-vol
-          emptyDir: {}
-  volumeClaimTemplates:
-    - metadata:
-        name: noobaastorage
-        labels:
-          app: noobaa
-      spec:
-        accessModes: ["ReadWriteOnce"]
-        resources:
-          requests:
-            storage: 30Gi
+  pvPool:
+    numVolumes: 3
+    resources:
+      requests:
+        storage: 32Gi
+    storageClass: STORAGE-CLASS-NAME
+  type: pv-pool
 ```
