+++ pkg/bucket/bucket.go
@@ -20,6 +20,7 @@ func Cmd() *cobra.Command {
 	cmd.AddCommand(
 		CmdCreate(),
 		CmdDelete(),
+		CmdStatus(),
 		CmdList(),
 	)
 	return cmd
@@ -41,6 +42,16 @@ func CmdDelete() *cobra.Command {
 		Use:   "delete <bucket-name>",
 		Short: "Delete a NooBaa bucket",
 		Run:   RunDelete,
+	}
+	return cmd
+}
+
+// CmdStatus returns a CLI command
+func CmdStatus() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "status <bucket-name>",
+		Short: "Show the status of a NooBaa bucket",
+		Run:   RunStatus,
 	}
 	return cmd
 }
@@ -83,6 +94,46 @@ func RunDelete(cmd *cobra.Command, args []string) {
 	}
 }
 
+// RunStatus runs a CLI command
+func RunStatus(cmd *cobra.Command, args []string) {
+	log := util.Logger()
+	if len(args) != 1 || args[0] == "" {
+		log.Fatalf(`Missing expected arguments: <bucket-name> %s`, cmd.UsageString())
+	}
+	bucketName := args[0]
+	nbClient := system.GetNBClient()
+	b, err := nbClient.ReadBucketAPI(nb.ReadBucketParams{Name: bucketName})
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	fmt.Printf("\n")
+	fmt.Printf("Bucket status:\n")
+	fmt.Printf("  %-22s : %s\n", "Bucket", b.Name)
+	if b.BucketClaim != nil {
+		fmt.Printf("  %-22s : %s\n", "OBC Namespace", b.BucketClaim.Namespace)
+		fmt.Printf("  %-22s : %s\n", "OBC BucketClass", b.BucketClaim.BucketClass)
+	}
+	fmt.Printf("  %-22s : %s\n", "Type", b.BucketType)
+	fmt.Printf("  %-22s : %s\n", "Mode", b.Mode)
+	if b.PolicyModes != nil {
+		fmt.Printf("  %-22s : %s\n", "ResiliencyStatus", b.PolicyModes.ResiliencyStatus)
+		fmt.Printf("  %-22s : %s\n", "QuotaStatus", b.PolicyModes.QuotaStatus)
+	}
+	if b.Undeletable != "" {
+		fmt.Printf("  %-22s : %s\n", "Undeletable", b.Undeletable)
+	}
+	if b.NumObjects != nil {
+		fmt.Printf("  %-22s : %d\n", "Num Objects", b.NumObjects.Value)
+	}
+	if b.DataCapacity != nil {
+		fmt.Printf("  %-22s : %s\n", "Data Size", nb.BigIntToHumanBytes(b.DataCapacity.Size))
+		fmt.Printf("  %-22s : %s\n", "Data Size Reduced", nb.BigIntToHumanBytes(b.DataCapacity.SizeReduced))
+		fmt.Printf("  %-22s : %s\n", "Data Space Avail", nb.BigIntToHumanBytes(b.DataCapacity.AvailableToUpload))
+	}
+	fmt.Printf("\n")
+}
+
 // RunList runs a CLI command
 func RunList(cmd *cobra.Command, args []string) {
 	nbClient := system.GetNBClient()
@@ -99,5 +150,7 @@ func RunList(cmd *cobra.Command, args []string) {
 		b := &list.Buckets[i]
 		table.AddRow(b.Name)
 	}
+	fmt.Printf("\n")
 	fmt.Print(table.String())
+	fmt.Printf("\n")
 }
